\name{XRcppExamples-package}
\alias{XRcppExamples-package}
\alias{XRcppExamples}
\docType{package}
\title{
Dummy package part of Rcpp unit testing
}
\description{
Dummy package part of Rcpp unit testing
}
\details{
\tabular{ll}{
Package: \tab XRcppExamples\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-09-06\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\keyword{ package }

